chore(arcade-staking-tvl): pr comments progress
