+++ projects/arcade-xyz/staking.test.js
@@ -1,11 +1,11 @@
-const { stakingTvl } = require('./staking_queries.js');
-const { CHAIN } = require('./staking_constants');
+const { stakingTvl } = require('./staking-queries.js');
+const { CHAIN } = require('./constants.js');
 
 const timestamp = Math.floor(Date.now() / 1000);
 const block = 19691786;
 const chainBlocks = { [CHAIN]: block };
 
-// to run: node projects/arcade-xyz/test.js
+// to run: node projects/arcade-xyz/staking.test.js
 
 async function main() {
   try {
