--- projects/arcade-xyz/staking_abi_single_sided.json
@@ -1,15 +1,0 @@
-[
-    {
-        "type": "function",
-        "name": "totalSupply",
-        "inputs": [],
-        "outputs": [
-        {
-            "name": "",
-            "type": "uint256",
-            "internalType": "uint256"
-        }
-        ],
-        "stateMutability": "view"
-   }
-]
